pyproject.toml
[tool.poetry]
name = "delphi"
version = "0.1.4"
description = "Forecasting library used at Maersk Broker. Based on DARTS."
authors = ["Raphael Daveau <rda@maerskbroker.com>"]
license = "UNLICENSED"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
"ruamel.yaml" = "^0.17.20"
numpy = "^1.23.0"
pandas = "^1.5.3"
kaleido = "^0.2.0"
plotly = "^5.13.0"
click = "^8.1.0"
joblib = "^1.2.0"
scikit-learn = "^1.2.0"
torch = "^2.0.0"
# working but only for py3.10
# torch = { url = "https://download.pytorch.org/whl/cu118/torch-2.0.0%2Bcu118-cp310-cp310-win_amd64.whl" }
pytorch-lightning = "^2.0.0"
# pytorch-forecasting = "^1.0.0"
darts = "^0.24.0"
optuna = "^3.1.0"
optuna-dashboard = "^0.9.0"
mlflow = "^2.7.0"
# [tool.poetry.dev-dependencies]
notebook = "^6.5.0"
ipywidgets = "^8.0.0"
mypy = "^1.1.0"
flake8 = "^3.9"
black = "^23.0.0"
pytest = "^6.0"
pytest-cov = "^2.0"
isort = "^5.11.1"

[tool.poetry.package]
include = ["py.typed"]

[tool.pytest.ini_options]
addopts = "--cov=src"
testpaths = ["tests"]

[tool.mypy]
mypy_path = "src"
check_untyped_defs = true
disallow_any_generics = true
ignore_missing_imports = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unreachable = true
warn_unused_configs = true
no_implicit_reexport = true

[tool.black]
line-length = 99
extend-exclude = ".git, docs/, *.csv, *.parquet, *.xslx"

[tool.isort]
profile = 'black'
src_paths = ['src', 'tests']
line_length = 99
lines_after_imports = 2
force_sort_within_sections = 'True'
order_by_type = 'True'

[tool.flake8]
max-line-length = 99
ignore = "E203, E266, E501, W503"
exclude = ".git, docs/, *.csv, *.parquet, *.xslx"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"